+++ src/com/codecool/termlib/gridelements/GridElement.java
@@ -1,0 +1,8 @@
+package com.codecool.termlib.gridelements;
+
+public class GridElement {
+    /**
+     * Character codes for the pipe elements of same shape
+     */
+    char charCode;
+}
